requirements.txt
mysql-connector-python==8.0.33
psycopg2==2.9.6
